require_relative for burgers_controller added to app.rb
Added index.erb for burgers.  Has a select for eatery and based on the
choice displays the relevant menu.

Added burgers_controller.rb and route for index.  The index route makes
all eateries available to the view to allow selection of eatery on which
to view burgers as burgers don't exist independently of an eatery.