Added new and create actions to burgers_controller.
Added small form and submit to burgers/index.erb to navigate to
burgers/new.  Also puts eatery_id into the params to use in the form for
burger creation.

Added burgers/new.erb with form for adding a new burger.  Puts eatery_id
into params as a burger needs to have an eatery.